chore: 랜덤 값 테스트 호출 장소 변경
